==> homework/02_project_setup.ipynb <==
# Environment & Config Check

from dotenv import load_dotenv
import os

load_dotenv()
print("API_KEY present:", os.getenv("API_KEY") is not None)


import numpy as np

arr = np.array([1, 2, 3])
print("Array sum:", np.sum(arr))
